requirements.txt
dj-database-url==0.5.0
Django==2.2.5
django-ckeditor==5.7.1
django-js-asset==1.2.2
Pillow==6.1.0
python-decouple==3.1
pytz==2019.2
sqlparse==0.3.0
